Re-organize s3 control & s3 data clients.
