+++ packages/storage/__tests__/providers/s3/apis/copy.test.ts
@@ -6,7 +6,7 @@ import { Amplify, StorageAccessLevel } from '@aws-amplify/core';
 
 import { StorageError } from '../../../../src/errors/StorageError';
 import { StorageValidationErrorCode } from '../../../../src/errors/types/validation';
-import { copyObject } from '../../../../src/providers/s3/utils/client';
+import { copyObject } from '../../../../src/providers/s3/utils/client/s3data';
 import { copy } from '../../../../src/providers/s3/apis';
 import {
 	CopyInput,
@@ -16,7 +16,7 @@ import {
 } from '../../../../src/providers/s3/types';
 import './testUtils';
 
-jest.mock('../../../../src/providers/s3/utils/client');
+jest.mock('../../../../src/providers/s3/utils/client/s3data');
 jest.mock('@aws-amplify/core', () => ({
 	ConsoleLogger: jest.fn().mockImplementation(function ConsoleLogger() {
 		return { debug: jest.fn() };
