+++ packages/storage/__tests__/providers/s3/apis/getUrl.test.ts
@@ -8,7 +8,7 @@ import { getUrl } from '../../../../src/providers/s3/apis';
 import {
 	getPresignedGetObjectUrl,
 	headObject,
-} from '../../../../src/providers/s3/utils/client';
+} from '../../../../src/providers/s3/utils/client/s3data';
 import {
 	GetUrlInput,
 	GetUrlOutput,
@@ -17,7 +17,7 @@ import {
 } from '../../../../src/providers/s3/types';
 import './testUtils';
 
-jest.mock('../../../../src/providers/s3/utils/client');
+jest.mock('../../../../src/providers/s3/utils/client/s3data');
 jest.mock('@aws-amplify/core', () => ({
 	ConsoleLogger: jest.fn().mockImplementation(function ConsoleLogger() {
 		return { debug: jest.fn() };
