+++ packages/storage/__tests__/providers/s3/apis/uploadData/putObjectJob.test.ts
@@ -4,12 +4,12 @@
 import { AWSCredentials } from '@aws-amplify/core/internals/utils';
 import { Amplify } from '@aws-amplify/core';
 
-import { putObject } from '../../../../../src/providers/s3/utils/client';
+import { putObject } from '../../../../../src/providers/s3/utils/client/s3data';
 import { calculateContentMd5 } from '../../../../../src/providers/s3/utils';
 import { putObjectJob } from '../../../../../src/providers/s3/apis/uploadData/putObjectJob';
 import '../testUtils';
 
-jest.mock('../../../../../src/providers/s3/utils/client');
+jest.mock('../../../../../src/providers/s3/utils/client/s3data');
 jest.mock('../../../../../src/providers/s3/utils', () => {
 	const utils = jest.requireActual('../../../../../src/providers/s3/utils');
