+++ packages/storage/__tests__/providers/s3/utils/client/S3/cases/deleteObject.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { deleteObject } from '../../../../../../../src/providers/s3/utils/client';
+import { deleteObject } from '../../../../../../../src/providers/s3/utils/client/s3data';
 import { ApiFunctionalTestCase } from '../../testUtils/types';
 
 import {
