+++ packages/storage/src/providers/s3/apis/internal/getUrl.ts
@@ -11,7 +11,7 @@ import {
 	GetUrlWithPathOutput,
 } from '../../types';
 import { StorageValidationErrorCode } from '../../../../errors/types/validation';
-import { getPresignedGetObjectUrl } from '../../utils/client';
+import { getPresignedGetObjectUrl } from '../../utils/client/s3data';
 import {
 	resolveS3ConfigAndInput,
 	validateStorageOperationInput,
