+++ packages/storage/src/providers/s3/apis/internal/list.ts
@@ -25,11 +25,11 @@ import {
 	ListObjectsV2Input,
 	ListObjectsV2Output,
 	listObjectsV2,
-} from '../../utils/client';
+} from '../../utils/client/s3data';
 import { getStorageUserAgentValue } from '../../utils/userAgent';
 import { logger } from '../../../../utils';
 import { STORAGE_INPUT_PREFIX } from '../../utils/constants';
-import { CommonPrefix } from '../../utils/client/types';
+import { CommonPrefix } from '../../utils/client/s3data/types';
 
 const MAX_PAGE_SIZE = 1000;
