+++ packages/storage/src/providers/s3/apis/uploadData/multipart/initialUpload.ts
@@ -5,7 +5,7 @@ import { StorageAccessLevel } from '@aws-amplify/core';
 
 import { ResolvedS3Config } from '../../../types/options';
 import { StorageUploadDataPayload } from '../../../../../types';
-import { Part, createMultipartUpload } from '../../../utils/client';
+import { Part, createMultipartUpload } from '../../../utils/client/s3data';
 import { logger } from '../../../../../utils';
 
 import {
