+++ packages/storage/src/providers/s3/apis/uploadData/multipart/uploadHandlers.ts
@@ -26,7 +26,7 @@ import {
 	abortMultipartUpload,
 	completeMultipartUpload,
 	headObject,
-} from '../../../utils/client';
+} from '../../../utils/client/s3data';
 import { getStorageUserAgentValue } from '../../../utils/userAgent';
 import { logger } from '../../../../../utils';
