+++ packages/storage/src/providers/s3/utils/client/s3control/index.ts
@@ -1,0 +1,13 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export {
+	getDataAccess,
+	GetDataAccessInput,
+	GetDataAccessOutput,
+} from '../s3control/getDataAccess';
+export {
+	listCallerAccessGrants,
+	ListCallerAccessGrantsInput,
+	ListCallerAccessGrantsOutput,
+} from '../s3control/listCallerAccessGrants';
