+++ packages/storage/src/providers/s3/utils/client/s3data/base.ts
@@ -12,7 +12,7 @@ import {
 	jitteredBackoff,
 } from '@aws-amplify/core/internals/aws-client-utils';
 
-import { parseXmlError } from './utils';
+import { parseXmlError } from '../utils';
 
 const DOMAIN_PATTERN = /^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$/;
 const IP_ADDRESS_PATTERN = /(\d+\.){3}\d+/;
@@ -53,7 +53,7 @@ export type S3EndpointResolverOptions = EndpointResolverOptions & {
  */
 const endpointResolver = (
 	options: S3EndpointResolverOptions,
-	apiInput?: { Bucket?: string; AccountId?: string },
+	apiInput?: { Bucket?: string },
 ) => {
 	const { region, useAccelerateEndpoint, customEndpoint, forcePathStyle } =
 		options;
@@ -61,11 +61,6 @@ const endpointResolver = (
 	// 1. get base endpoint
 	if (customEndpoint) {
 		endpoint = new AmplifyUrl(customEndpoint);
-	} else if (apiInput?.AccountId) {
-		// Control plane operations
-		endpoint = new AmplifyUrl(
-			`https://s3-control.${region}.${getDnsSuffix(region)}`,
-		);
 	} else if (useAccelerateEndpoint) {
 		if (forcePathStyle) {
 			throw new Error(
@@ -77,7 +72,7 @@ const endpointResolver = (
 		endpoint = new AmplifyUrl(`https://s3.${region}.${getDnsSuffix(region)}`);
 	}
 	// 2. inject bucket name
-	if (apiInput?.Bucket && !apiInput.AccountId) {
+	if (apiInput?.Bucket) {
 		if (!isDnsCompatibleBucketName(apiInput.Bucket)) {
 			throw new Error(`Invalid bucket name: "${apiInput.Bucket}".`);
 		}
