+++ packages/storage/src/providers/s3/utils/client/s3data/copyObject.ts
@@ -10,8 +10,6 @@ import {
 import { AmplifyUrl } from '@aws-amplify/core/internals/utils';
 import { composeServiceApi } from '@aws-amplify/core/internals/aws-client-utils/composers';
 
-import type { CopyObjectCommandInput, CopyObjectCommandOutput } from './types';
-import { defaultConfig } from './base';
 import {
 	assignStringVariables,
 	buildStorageServiceError,
@@ -21,7 +19,10 @@ import {
 	serializeObjectConfigsToHeaders,
 	serializePathnameObjectKey,
 	validateS3RequiredParameter,
-} from './utils';
+} from '../utils';
+
+import type { CopyObjectCommandInput, CopyObjectCommandOutput } from './types';
+import { defaultConfig } from './base';
 
 export type CopyObjectInput = Pick<
 	CopyObjectCommandInput,
