+++ packages/storage/src/providers/s3/utils/client/s3data/listObjectsV2.ts
@@ -13,11 +13,6 @@ import {
 } from '@aws-amplify/core/internals/utils';
 import { composeServiceApi } from '@aws-amplify/core/internals/aws-client-utils/composers';
 
-import type {
-	ListObjectsV2CommandInput,
-	ListObjectsV2CommandOutput,
-} from './types';
-import { defaultConfig } from './base';
 import {
 	assignStringVariables,
 	buildStorageServiceError,
@@ -29,7 +24,13 @@ import {
 	parseXmlBody,
 	parseXmlError,
 	s3TransferHandler,
-} from './utils';
+} from '../utils';
+
+import type {
+	ListObjectsV2CommandInput,
+	ListObjectsV2CommandOutput,
+} from './types';
+import { defaultConfig } from './base';
 
 export type ListObjectsV2Input = ListObjectsV2CommandInput;
