+++ packages/storage/src/providers/s3/utils/client/s3data/putObject.ts
@@ -10,77 +10,86 @@ import {
 import { AmplifyUrl } from '@aws-amplify/core/internals/utils';
 import { composeServiceApi } from '@aws-amplify/core/internals/aws-client-utils/composers';
 
-import { defaultConfig } from './base';
-import type { HeadObjectCommandInput, HeadObjectCommandOutput } from './types';
 import {
+	assignStringVariables,
 	buildStorageServiceError,
-	deserializeMetadata,
-	deserializeNumber,
-	deserializeTimestamp,
 	map,
 	parseXmlError,
 	s3TransferHandler,
+	serializeObjectConfigsToHeaders,
 	serializePathnameObjectKey,
 	validateS3RequiredParameter,
-} from './utils';
+} from '../utils';
 
-export type HeadObjectInput = Pick<HeadObjectCommandInput, 'Bucket' | 'Key'>;
+import { defaultConfig } from './base';
+import type { PutObjectCommandInput, PutObjectCommandOutput } from './types';
 
-export type HeadObjectOutput = Pick<
-	HeadObjectCommandOutput,
-	| 'ContentLength'
+export type PutObjectInput = Pick<
+	PutObjectCommandInput,
+	| 'Bucket'
+	| 'Key'
+	| 'Body'
+	| 'ACL'
+	| 'CacheControl'
+	| 'ContentDisposition'
+	| 'ContentEncoding'
 	| 'ContentType'
-	| 'ETag'
-	| 'LastModified'
+	| 'ContentMD5'
+	| 'Expires'
 	| 'Metadata'
-	| 'VersionId'
-	| '$metadata'
+	| 'Tagging'
 >;
 
-const headObjectSerializer = async (
-	input: HeadObjectInput,
+export type PutObjectOutput = Pick<
+	PutObjectCommandOutput,
+	// PutObject output is not exposed in public API, but only logged in the debug mode
+	// so we only expose $metadata, ETag and VersionId for debug purpose.
+	'$metadata' | 'ETag' | 'VersionId'
+>;
+
+const putObjectSerializer = async (
+	input: PutObjectInput,
 	endpoint: Endpoint,
 ): Promise<HttpRequest> => {
+	const headers = {
+		...(await serializeObjectConfigsToHeaders({
+			...input,
+			ContentType: input.ContentType ?? 'application/octet-stream',
+		})),
+		...assignStringVariables({ 'content-md5': input.ContentMD5 }),
+	};
 	const url = new AmplifyUrl(endpoint.url.toString());
 	validateS3RequiredParameter(!!input.Key, 'Key');
 	url.pathname = serializePathnameObjectKey(url, input.Key);
 
 	return {
-		method: 'HEAD',
-		headers: {},
+		method: 'PUT',
+		headers,
 		url,
+		body: input.Body,
 	};
 };
 
-const headObjectDeserializer = async (
+const putObjectDeserializer = async (
 	response: HttpResponse,
-): Promise<HeadObjectOutput> => {
+): Promise<PutObjectOutput> => {
 	if (response.statusCode >= 300) {
-		// error is always set when statusCode >= 300
 		const error = (await parseXmlError(response)) as Error;
 		throw buildStorageServiceError(error, response.statusCode);
 	} else {
-		const contents = {
+		return {
 			...map(response.headers, {
-				ContentLength: ['content-length', deserializeNumber],
-				ContentType: 'content-type',
 				ETag: 'etag',
-				LastModified: ['last-modified', deserializeTimestamp],
 				VersionId: 'x-amz-version-id',
 			}),
-			Metadata: deserializeMetadata(response.headers),
-		};
-
-		return {
 			$metadata: parseMetadata(response),
-			...contents,
 		};
 	}
 };
 
-export const headObject = composeServiceApi(
+export const putObject = composeServiceApi(
 	s3TransferHandler,
-	headObjectSerializer,
-	headObjectDeserializer,
+	putObjectSerializer,
+	putObjectDeserializer,
 	{ ...defaultConfig, responseType: 'text' },
 );
