+++ packages/storage/src/providers/s3/utils/client/s3data/uploadPart.ts
@@ -7,59 +7,54 @@ import {
 	HttpResponse,
 	parseMetadata,
 } from '@aws-amplify/core/internals/aws-client-utils';
-import { AmplifyUrl } from '@aws-amplify/core/internals/utils';
+import {
+	AmplifyUrl,
+	AmplifyUrlSearchParams,
+} from '@aws-amplify/core/internals/utils';
 import { composeServiceApi } from '@aws-amplify/core/internals/aws-client-utils/composers';
 
-import { defaultConfig } from './base';
-import type { PutObjectCommandInput, PutObjectCommandOutput } from './types';
 import {
 	assignStringVariables,
 	buildStorageServiceError,
 	map,
 	parseXmlError,
 	s3TransferHandler,
-	serializeObjectConfigsToHeaders,
 	serializePathnameObjectKey,
 	validateS3RequiredParameter,
-} from './utils';
+} from '../utils';
 
-export type PutObjectInput = Pick<
-	PutObjectCommandInput,
-	| 'Bucket'
-	| 'Key'
-	| 'Body'
-	| 'ACL'
-	| 'CacheControl'
-	| 'ContentDisposition'
-	| 'ContentEncoding'
-	| 'ContentType'
-	| 'ContentMD5'
-	| 'Expires'
-	| 'Metadata'
-	| 'Tagging'
+import { defaultConfig } from './base';
+import type { UploadPartCommandInput, UploadPartCommandOutput } from './types';
+
+// Content-length is ignored here because it's forbidden header
+// and will be set by browser or fetch polyfill.
+export type UploadPartInput = Pick<
+	UploadPartCommandInput,
+	'PartNumber' | 'Body' | 'UploadId' | 'Bucket' | 'Key' | 'ContentMD5'
 >;
 
-export type PutObjectOutput = Pick<
-	PutObjectCommandOutput,
-	// PutObject output is not exposed in public API, but only logged in the debug mode
-	// so we only expose $metadata, ETag and VersionId for debug purpose.
-	'$metadata' | 'ETag' | 'VersionId'
+export type UploadPartOutput = Pick<
+	UploadPartCommandOutput,
+	'$metadata' | 'ETag'
 >;
 
-const putObjectSerializer = async (
-	input: PutObjectInput,
+const uploadPartSerializer = async (
+	input: UploadPartInput,
 	endpoint: Endpoint,
 ): Promise<HttpRequest> => {
 	const headers = {
-		...(await serializeObjectConfigsToHeaders({
-			...input,
-			ContentType: input.ContentType ?? 'application/octet-stream',
-		})),
 		...assignStringVariables({ 'content-md5': input.ContentMD5 }),
 	};
+	headers['content-type'] = 'application/octet-stream';
 	const url = new AmplifyUrl(endpoint.url.toString());
 	validateS3RequiredParameter(!!input.Key, 'Key');
 	url.pathname = serializePathnameObjectKey(url, input.Key);
+	validateS3RequiredParameter(!!input.PartNumber, 'PartNumber');
+	validateS3RequiredParameter(!!input.UploadId, 'UploadId');
+	url.search = new AmplifyUrlSearchParams({
+		partNumber: input.PartNumber + '',
+		uploadId: input.UploadId,
+	}).toString();
 
 	return {
 		method: 'PUT',
@@ -69,9 +64,9 @@ const putObjectSerializer = async (
 	};
 };
 
-const putObjectDeserializer = async (
+const uploadPartDeserializer = async (
 	response: HttpResponse,
-): Promise<PutObjectOutput> => {
+): Promise<UploadPartOutput> => {
 	if (response.statusCode >= 300) {
 		const error = (await parseXmlError(response)) as Error;
 		throw buildStorageServiceError(error, response.statusCode);
@@ -79,16 +74,15 @@ const putObjectDeserializer = async (
 		return {
 			...map(response.headers, {
 				ETag: 'etag',
-				VersionId: 'x-amz-version-id',
 			}),
 			$metadata: parseMetadata(response),
 		};
 	}
 };
 
-export const putObject = composeServiceApi(
+export const uploadPart = composeServiceApi(
 	s3TransferHandler,
-	putObjectSerializer,
-	putObjectDeserializer,
+	uploadPartSerializer,
+	uploadPartDeserializer,
 	{ ...defaultConfig, responseType: 'text' },
 );
