+++ scripts/dts-bundler/dts-bundler.config.js
@@ -74,7 +74,7 @@ const config = {
 		},
 		{
 			filePath: './s3.d.ts',
-			outFile: join(storagePackageSrcClientsPath, 'client', 'types.ts'),
+			outFile: join(storagePackageSrcClientsPath, 'client', 's3data', 'types.ts'),
 			libraries: {
 				inlinedLibraries: ['@aws-sdk/client-s3'],
 			},
